// src/types/style.types.ts
export type TSize = "tiny" | "small" | "medium" | "large" | "giant";

export type TStatus =
  | "basic"
  | "primary"
  | "success"
  | "info"
  | "warning"
  | "danger"
  | "control";

export type TAppearance = "filled" | "outline" | "ghost";

export type TShape = "rectangle" | "semi-round" | "round";

export type TFieldSize = "small" | "medium" | "large";